Add hub_name
