+++ paddlehub/commands/config.py
@@ -25,7 +25,6 @@ import re
 from paddlehub.commands.base_command import BaseCommand, ENTRY
 from paddlehub.common.dir import CONF_HOME
 from paddlehub.common.server_config import default_server_config
-from paddlehub.common.hub_server import CacheUpdater
 from paddlehub.common.hub_server import HubServer
 
 HubServer()
@@ -102,7 +101,6 @@ class ConfigCommand(BaseCommand):
         print(str)
 
     def execute(self, argv):
-        CacheUpdater("hub_config").start()
         args = self.parser.parse_args()
         if args.option is None:
             ConfigCommand.show_config()
