+++ paddlehub/common/srv_utils.py
@@ -15,8 +15,15 @@
 import requests
 import paddle
 import json
+import time
+import uuid
+import os
 
 from paddlehub import version
+from paddlehub.common.dir import CONF_HOME
+from paddlehub.common.decorator_utils import singleton
+from paddlehub.common.utils import md5
+from paddlehub.common.server_config import default_server_config
 
 
 def uri_path(server_url, api):
@@ -37,3 +44,24 @@ def hub_request(api, params, extra=None, timeout=8):
     params["extra"] = json.dumps(extra)
     r = requests.get(api, params, timeout=timeout)
     return r.json()
+
+
+@singleton
+class ConfigInfo(object):
+    def __init__(self):
+        self.filepath = os.path.join(CONF_HOME, "config.json")
+        self.hub_name = None
+        self.configs = None
+        if os.path.exists(self.filepath):
+            with open(self.filepath, "r") as fp:
+                self.configs = json.load(fp)
+                self.hub_name = self.configs.get("hub_name", None)
+
+    def get_hub_name(self):
+        if self.hub_name is None:
+            self.hub_name = md5(str(uuid.uuid1())[-12:]) + "-" + str(
+                int(time.time()))
+            with open(self.filepath, "w") as fp:
+                fp.write(json.dumps(default_server_config))
+
+        return self.hub_name
